requirements.txt
anyio==4.9.0
certifi==2025.1.31
exceptiongroup==1.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-telegram-bot==22.0
sniffio==1.3.1
SQLAlchemy==2.0.39
typing_extensions==4.12.2
